requirements.txt
Flask==3.1.0
Flask-CORS==5.0.1
python-docx==1.1.2
python-pptx==1.0.2
pandas==2.2.3
python-dotenv==1.0.1
groq==0.18.0
aiohttp==3.11.13
lxml==5.3.1
typing-extensions==4.12.2
